Invitacion

+++ app/src/main/java/com/example/gestionpisoscompartidos/model/dtos/PointDeltaDTO.kt
@@ -1,8 +1,9 @@
 package com.example.gestionpisoscompartidos.model.dtos
 
 data class PointDeltaDTO(
-    val x: Short,
-    val y: Short,
+    // x e y estaban en Short, pero en pizarraView meten float como parametros
+    val x: Float,
+    val y: Float,
     val size: Float,
     var color: Byte,
 )
